// front-mobile/types/enums.ts
export enum ShoppingListStatus {
  PENDING = "pending",
  COMPLETED = "completed",
  CANCELLED = "cancelled",
}

export enum ShoppingListItemStatus {
  PENDING = "pending",
  BOUGHT = "bought",
  CANCELLED = "cancelled",
}

export enum Unit {
  GRAM = "g",
  KILOGRAM = "kg",
  LITER = "L",
  MILLILITER = "mL",
  PIECE = "piece",
  PACK = "pack",
  BOTTLE = "bottle",
  BOX = "box",
  BAG = "bag",
  CAN = "can",
  JAR = "jar",
  BUNCH = "bunch",
  HEAD = "head",
  CLOVE = "clove",
  LEAF = "leaf",
  SLICE = "slice",
  CUP = "cup",
  TABLESPOON = "tablespoon",
  TEASPOON = "teaspoon",
  PINCH = "pinch",
  DASH = "dash",
}